+++ private_source_manager/app/psm/__init__.py
@@ -1,3 +1,3 @@
 """Private Source Manager add-on."""
 
-__version__ = "0.1.1"
+__version__ = "0.1.2"